[modify] slot design can be defined outside
