+++ src/app/components/Timetable/slots/daySlot.tsx
@@ -1,11 +1,13 @@
 import { Day } from "@/app/type";
 import { SlotDiv, slotProps } from "./slot";
+import React from "react";
 
 /**
  * 曜日が表示されるスロット要素のプロパティ
  */
 interface dayProps extends slotProps {
   day: Day;
+  daySlotElement: React.FC<Day>;
 }
 
 /**
@@ -14,5 +16,9 @@ interface dayProps extends slotProps {
  * @returns 曜日スロット要素
  */
 export const DaySlot: React.FC<dayProps> = (props: dayProps) => {
-  return <SlotDiv className="col-span-2 row-span-1">{props.day}</SlotDiv>;
+  return (
+    <SlotDiv className="col-span-2 row-span-1">
+      {props.daySlotElement(props.day)}
+    </SlotDiv>
+  );
 };
