+++ src/app/components/Timetable/slots/periodSlot.tsx
@@ -6,6 +6,7 @@ import { Period, SlotDiv, slotProps } from "./slot";
  */
 interface periodProps extends slotProps {
   period: Period;
+  periodSlotElement: React.FC<Period>;
 }
 
 /**
@@ -14,5 +15,9 @@ interface periodProps extends slotProps {
  * @returns 時限スロット要素
  */
 export const PeriodSlot: React.FC<periodProps> = (props: periodProps) => {
-  return <SlotDiv className="col-span-1 row-span-2">{props.period}</SlotDiv>;
+  return (
+    <SlotDiv className="col-span-1 row-span-2">
+      {props.periodSlotElement(props.period)}
+    </SlotDiv>
+  );
 };
